refactor(actions): notes/ como directorio con barril
