+++ src/actions/notes/getNotes.ts
@@ -1,0 +1,14 @@
+'use server'
+
+import { prisma } from '@/lib/prisma'
+import type { Notes } from '@/modules/notes/model/notes'
+import { requireBoardAccess } from '../shared'
+
+export async function getNotes({ boardId }: { boardId: string }): Promise<Notes> {
+	await requireBoardAccess(boardId)
+
+	const note = await prisma.note.findUnique({
+		where: { boardId },
+	})
+	return note?.content ?? ''
+}
