+++ src/actions/notes/index.ts
@@ -1,0 +1,2 @@
+export * from './getNotes'
+export * from './saveNotes'
